+++ ext/or-tools/constraint.cpp
@@ -151,14 +151,26 @@ BoolVarSpan from_ruby<BoolVarSpan>(Object x)
   return BoolVarSpan(x);
 }
 
+Rice::Class rb_cSatIntVar;
+
 void init_constraint(Rice::Module& m) {
-  Rice::define_class_under<IntVar>(m, "SatIntVar")
+  rb_cSatIntVar = Rice::define_class_under<IntVar>(m, "SatIntVar")
     .define_method("name", &IntVar::Name);
 
   Rice::define_class_under<IntervalVar>(m, "SatIntervalVar")
     .define_method("name", &IntervalVar::Name);
 
-  Rice::define_class_under<Constraint>(m, "SatConstraint");
+  Rice::define_class_under<Constraint>(m, "SatConstraint")
+    .define_method(
+      "only_enforce_if",
+      *[](Constraint& self, Object literal) {
+        if (literal.is_a(rb_cSatIntVar)) {
+          return self.OnlyEnforceIf(from_ruby<IntVar>(literal).ToBoolVar());
+        } else {
+          // TODO support BoolVarSpan
+          return self.OnlyEnforceIf(from_ruby<BoolVar>(literal));
+        }
+      });
 
   Rice::define_class_under<BoolVar>(m, "BoolVar")
     .define_method("name", &BoolVar::Name)
@@ -219,113 +231,113 @@ void init_constraint(Rice::Module& m) {
     .define_method(
       "add_bool_or",
       *[](CpModelBuilder& self, BoolVarSpan literals) {
-        self.AddBoolOr(literals);
+        return self.AddBoolOr(literals);
       })
     .define_method(
       "add_bool_and",
       *[](CpModelBuilder& self, BoolVarSpan literals) {
-        self.AddBoolAnd(literals);
+        return self.AddBoolAnd(literals);
       })
     .define_method(
       "add_bool_xor",
       *[](CpModelBuilder& self, BoolVarSpan literals) {
-        self.AddBoolXor(literals);
+        return self.AddBoolXor(literals);
       })
     .define_method(
       "add_implication",
       *[](CpModelBuilder& self, BoolVar a, BoolVar b) {
-        self.AddImplication(a, b);
+        return self.AddImplication(a, b);
       })
     .define_method(
       "add_equality",
       *[](CpModelBuilder& self, LinearExpr x, LinearExpr y) {
-        self.AddEquality(x, y);
+        return self.AddEquality(x, y);
       })
     .define_method(
       "add_greater_or_equal",
       *[](CpModelBuilder& self, LinearExpr x, LinearExpr y) {
-        self.AddGreaterOrEqual(x, y);
+        return self.AddGreaterOrEqual(x, y);
       })
     .define_method(
       "add_greater_than",
       *[](CpModelBuilder& self, LinearExpr x, LinearExpr y) {
-        self.AddGreaterThan(x, y);
+        return self.AddGreaterThan(x, y);
       })
     .define_method(
       "add_less_or_equal",
       *[](CpModelBuilder& self, LinearExpr x, LinearExpr y) {
-        self.AddLessOrEqual(x, y);
+        return self.AddLessOrEqual(x, y);
       })
     .define_method(
       "add_less_than",
       *[](CpModelBuilder& self, LinearExpr x, LinearExpr y) {
-        self.AddLessThan(x, y);
+        return self.AddLessThan(x, y);
       })
     // TODO add domain
     // .define_method(
     //   "add_linear_constraint",
     //   *[](CpModelBuilder& self, LinearExpr expr, Domain domain) {
-    //     self.AddLinearConstraint(expr, domain);
+    //     return self.AddLinearConstraint(expr, domain);
     //   })
     .define_method(
       "add_not_equal",
       *[](CpModelBuilder& self, LinearExpr x, LinearExpr y) {
-        self.AddNotEqual(x, y);
+        return self.AddNotEqual(x, y);
       })
     .define_method(
       "add_all_different",
       *[](CpModelBuilder& self, IntVarSpan vars) {
-        self.AddAllDifferent(vars);
+        return self.AddAllDifferent(vars);
       })
     .define_method(
       "add_inverse_constraint",
       *[](CpModelBuilder& self, IntVarSpan variables, IntVarSpan inverse_variables) {
-        self.AddInverseConstraint(variables, inverse_variables);
+        return self.AddInverseConstraint(variables, inverse_variables);
       })
     .define_method(
       "add_min_equality",
       *[](CpModelBuilder& self, IntVar target, IntVarSpan vars) {
-        self.AddMinEquality(target, vars);
+        return self.AddMinEquality(target, vars);
       })
     .define_method(
       "add_lin_min_equality",
       *[](CpModelBuilder& self, LinearExpr target, LinearExprSpan exprs) {
-        self.AddLinMinEquality(target, exprs);
+        return self.AddLinMinEquality(target, exprs);
       })
     .define_method(
       "add_max_equality",
       *[](CpModelBuilder& self, IntVar target, IntVarSpan vars) {
-        self.AddMaxEquality(target, vars);
+        return self.AddMaxEquality(target, vars);
       })
     .define_method(
       "add_lin_max_equality",
       *[](CpModelBuilder& self, LinearExpr target, LinearExprSpan exprs) {
-        self.AddLinMaxEquality(target, exprs);
+        return self.AddLinMaxEquality(target, exprs);
       })
     .define_method(
       "add_division_equality",
       *[](CpModelBuilder& self, IntVar target, IntVar numerator, IntVar denominator) {
-        self.AddDivisionEquality(target, numerator, denominator);
+        return self.AddDivisionEquality(target, numerator, denominator);
       })
     .define_method(
       "add_abs_equality",
       *[](CpModelBuilder& self, IntVar target, IntVar var) {
-        self.AddAbsEquality(target, var);
+        return self.AddAbsEquality(target, var);
       })
     .define_method(
       "add_modulo_equality",
       *[](CpModelBuilder& self, IntVar target, IntVar var, IntVar mod) {
-        self.AddModuloEquality(target, var, mod);
+        return self.AddModuloEquality(target, var, mod);
       })
     .define_method(
       "add_product_equality",
       *[](CpModelBuilder& self, IntVar target, IntVarSpan vars) {
-        self.AddProductEquality(target, vars);
+        return self.AddProductEquality(target, vars);
       })
     .define_method(
       "add_no_overlap",
       *[](CpModelBuilder& self, IntervalVarSpan vars) {
-        self.AddNoOverlap(vars);
+        return self.AddNoOverlap(vars);
       })
     .define_method(
       "maximize",