+++ test/expression_test.rb
@@ -39,6 +39,16 @@ class ExpressionTest < Minitest::Test
     assert_equal 0, solver.value(y)
   end
 
+  def test_only_enforce_if
+    model = ORTools::CpModel.new
+    x = model.new_int_var(0, 1, "x")
+    model.add(x == x + 1).only_enforce_if(x)
+
+    solver = ORTools::CpSolver.new
+    assert_equal :optimal, solver.solve(model)
+    assert_equal(0, solver.value(x))
+  end
+
   def test_to_s
     model = ORTools::CpModel.new
     x = model.new_int_var(0, 1, "x")